basic html + started off js file
Created a basic HTML file with the command lines to use on the app. Linked the JS file.

In the JS file, created a variable called input to store the prompt asking what the user would like to do. Added that variable to a while loop, in order to have the message pop-up nonstop until the user input either 'quit' or 'q'. Added a console message to confirm it too.