implemented list
Made an array for the todos list, adding some values in strings. Listed the todos with a For Loop, displaying the index to make it easier for the user whenever choosing which task to remove, in case they input 'delete'.

In the while loop, created a if in case it's equal to 'list', with two console messages of dashes to separate the starter of the list in the console.

Using the For loop, starting from index 0, adding one each time to the total length. The message in the console is written as a string template literal, adding the 'i' as a reference to the index and then the todos with the index (i).