implemented 'new' and 'delete'
Added another conditional - else if. If it's equal to 'new', it will prompt the user again and ask what's the new todo. Using const because it's only used once, whilst let was previously used because it was in a forever loop.

Added the results of that prompt to the 'todos' array by using the array method 'todos.push'. Added a console message to give some feedback, using a string template literal.

-
Added yet another conditional, and a variable named 'index' with a prompt message asking for the index to be deleted. Created another variable called 'deleted' that uses the array method splice to delete from the index. Added a feedback message in the console, confirming action.

-
To prevent user from entering an invalid value, parsed the index string to turn the string into a number. Added it to a variable called index, and amended the index variable to indexString to make it more readable.

-
Added another conditional to validate the NaN and give a feedback message in the console confirming it's not a number.